restructure naming
