--- app/templates/check.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<pingdom_http_custom_check>
-<status>OK</status>
-<response_time>10823.74</response_time>
-</pingdom_http_custom_check>
