--- app/templates/logs.html
@@ -1,24 +1,0 @@
-<html>
-<body>
-<h1>Redirect check - logs</h1>
-<p>
-Source: http://www.great.gov.uk/de/invest    Expected result: https://invest.great.gov.uk/de    Actual result: https://invest.great.gov.uk/de -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export/opportunities    Expected result: https://www.exportingisgreat.gov.uk/opportunities    Actual result: https://www.exportingisgreat.gov.uk/opportunities -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export/selling-online-overseas    Expected result: https://selling-online-overseas.export.great.gov.uk/markets    Actual result: https://selling-online-overseas.export.great.gov.uk/markets -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/study    Expected result: https://study-uk.britishcouncil.org    Actual result: https://study-uk.britishcouncil.org -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export/regular    Expected result: https://www.exportingisgreat.gov.uk/regular    Actual result: https://www.exportingisgreat.gov.uk/regular -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export    Expected result: https://www.exportingisgreat.gov.uk    Actual result: https://www.exportingisgreat.gov.uk -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export/find-a-buyer    Expected result: https://find-a-buyer.export.great.gov.uk    Actual result: https://find-a-buyer.export.great.gov.uk -- Result: GOOD<br/>
-Source: http://export.great.gov.uk    Expected result: https://www.exportingisgreat.gov.uk    Actual result: https://www.exportingisgreat.gov.uk -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/int/invest    Expected result: https://invest.great.gov.uk/int    Actual result: https://invest.great.gov.uk/int -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/trade    Expected result: https://trade.great.gov.uk    Actual result: https://trade.great.gov.uk -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/invest    Expected result: https://invest.great.gov.uk    Actual result: https://invest.great.gov.uk -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/in/invest    Expected result: https://invest.great.gov.uk/in    Actual result: https://invest.great.gov.uk/in -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/us/invest    Expected result: https://invest.great.gov.uk/us    Actual result: https://invest.great.gov.uk/us -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export/new    Expected result: https://www.exportingisgreat.gov.uk/new    Actual result: https://www.exportingisgreat.gov.uk/new -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/export/occasional    Expected result: https://www.exportingisgreat.gov.uk/occasional    Actual result: https://www.exportingisgreat.gov.uk/occasional -- Result: GOOD<br/>
-Source: http://www.great.gov.uk/cn/invest    Expected result: https://invest.great.gov.uk/cn    Actual result: https://invest.great.gov.uk/cn -- Result: GOOD<br/>
---  All Sites Checked -- OK --
-</p>
-</body>
-</html>
