+++ check_redirects/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class CheckredirectsConfig(AppConfig):
-    name = 'checkredirects'
+    name = 'check_redirects'
