+++ check_redirects/urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-from checkredirects import views
+from check_redirects import views
 
 urlpatterns = [
     url(r'^$', views.url_search_results, name='url_search_result'),
