+++ check_redirects/views.py
