--- checkredirects/templates/refresh.html
@@ -1,9 +1,0 @@
-<html>
-<body>
-<h1>Refreshing</h1>
-<p>
-Check Logs for progress:
-<a href="logs.html">Progress Logs</a>
-</p>
-</body>
-</html>
