+++ config/sample_env
@@ -1,0 +1,3 @@
+DEBUG=on
+SECRET_KEY=do-not-use-in-prod
+DATABASE_URL=postgres://localhost/redirectcheckerdb
