+++ config/wsgi.py
@@ -13,6 +13,6 @@ from django.core.wsgi import get_wsgi_application
 from dotenv import load_dotenv
 load_dotenv()
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "redirectchecker.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings")
 
 application = get_wsgi_application()
